// src/types/api/Ownership.ts
/* eslint-disable */
/* @ts-ignore */
/**
 * DO NOT MODIFY IT BY HAND.
 * This file was automatically generated.
 */

import { Currency } from './Currency';

export interface Ownership {
  address: string;
  avatar?: string;
  currency: Currency;
  name?: string;
  price?: string;
  quantity?: string;
  sellingQuantity?: string;
  url?: string;
}